refactor: rename smtp client to smtp service
